requirements.txt
anyio==3.6.2
asttokens==2.4.1
certifi==2022.12.7
charset-normalizer==3.3.2
colorama==0.4.6
executing==2.0.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
icecream==2.1.3
idna==3.4
numpy==1.26.4
pandas==2.2.1
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-telegram-bot==20.2
pytz==2024.1
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tzdata==2024.1
urllib3==2.2.1
